==> code/4d_explanation_interpretml.ipynb <==
### Import necessary libraries

import pandas as pd
import numpy as np
from sklearn.ensemble import RandomForestClassifier
import matplotlib.pyplot as plt
import os
import pickle 

import interpret
import shap
from interpret.blackbox import ShapKernel, LimeTabular

path_list = ('results/', 'model_explanation/', 'interpretML')
cur_path = ''
for rep in path_list:
    cur_path += rep
    if not os.path.exists(cur_path):
        os.makedirs(cur_path)
result_path = ''.join(path_list)

# CREATE EXPLANATIONS FOR RANDOM FOREST 
**File**: 5_explanation_interpretml.ipynb

**Author**: Sebastian Benno Veuskens 

**Date**: 2024-07-28

**Data**: train and test (same as evaluated model)  


## Modify

**Working directory**

os.chdir("C:/Users/Sebastian's work/OneDrive - OptiMedis AG/Dokumente/Coding/High-Cost-patient-analysis")

**Parameters & Settings**

target = 'HC_Patient_Next_Year'
excluded = 'Total_Costs_Next_Year'
prediction_outcome = 1  # Indicate which prediction outcome the sample should have 
true_outcome = 1 # Indicate which true HCP status the patient in the sample should have

### Load data

train = pd.read_csv('data/train_validate.csv', sep=',', header=0)
test = pd.read_csv('data/test.csv', sep=',', header=0)

### Transform categorical values

# TODO: Check if I need this or whether I need to make the same transformations as before
# train['Sex'] = pd.factorize(train['Sex'])[0]
# test['Sex'] = pd.factorize(test['Sex'])[0]

from sklearn.preprocessing import OrdinalEncoder
all_columns = train.columns

# If a column has more than two values, it is not categorical
categorical_columns = all_columns[train.nunique(axis=0) == 2]
numerical_columns = [all_columns.difference(categorical_columns)]

categorical_columns_indices = [test.columns.tolist().index(cn) for cn in categorical_columns]
oe = OrdinalEncoder().fit(train[categorical_columns])
categorical_name_mapping = {i: list(v) if i < 11 else ['diagnosis absent', 'diagnosis present']
                            for (i, v) in zip(categorical_columns_indices, oe.categories_)}


train.loc[:, categorical_columns] = oe.transform(train[categorical_columns])
test.loc[:, categorical_columns] = oe.transform(test[categorical_columns])

### Split predictors & outcome labels 

predictors = [var for var in train.columns if var not in (target, excluded)]
X_train, y_train = train[predictors], train[target]
X_test, y_test = test[predictors], test[target]

train.head()

## Load Model

with open('results/model_explanation/random_forest_python.sav', 'rb') as handle:
    model = pickle.load(handle)
    
with open('results/model_explanation/random_forest_python_threshold.sav', 'rb') as handle:
    best_thresh = pickle.load(handle)

best_thresh

### Choose sample
Sample to investigate better, choose sample with user-specified model outcome

prediction_probs = model.predict_proba(X_test)[:,1]
prediction_probs_pos = prediction_probs[y_test == 1]
prediction_probs_neg = prediction_probs[y_test == 0]

predictions = np.array(prediction_probs >= best_thresh, dtype=int)
predictions_pos = np.array(prediction_probs_pos >= best_thresh, dtype=int)
predictions_neg = np.array(prediction_probs_neg >= best_thresh, dtype=int)

samples_selected = X_test.loc[(y_test == true_outcome) & (predictions == prediction_outcome)]
samples_true_pos = X_test.loc[(y_test == 1) & (predictions == 1)]
samples_false_pos = X_test.loc[(y_test == 0) & (predictions == 1)]
samples_false_neg = X_test.loc[(y_test == 1) & (predictions == 0)]
samples_true_neg = X_test.loc[(y_test == 0) & (predictions == 0)]

## XAI METHODS

### MODIFY

# Indicate the indices of the samples I would like to explain
# local_samples_idx = list(samples_true_pos.index[:2]) + list(samples_false_pos.index[:2]) + list(samples_false_neg.index[:2]) + list(samples_true_neg.index[:2])
local_samples_idx = [1467, 33839] 
local_samples_X = X_test.iloc[local_samples_idx,]
local_samples_y = y_test.iloc[local_samples_idx]
local_samples_predictions = predictions[local_samples_idx]

### SHAP 

a = 62159 + 5094
5094 / a

# Due to size of data set, summarization techniques are required
# TODO: Figure out a good summarize number here 
X_train_summary = shap.sample(X_train, 200)
exp_shap = ShapKernel(model, X_train_summary)

# shap = ShapKernel(model, X_train)
exp_local_shap = exp_shap.explain_local(local_samples_X, local_samples_y)

interpret.show(exp_local_shap)

### LIME 

exp_lime = LimeTabular(model, X_train)
exp_local_lime = exp_lime.explain_local(local_samples_X, local_samples_y)

# TODO: What does the 0 do here? 
interpret.show(exp_local_lime, 1)

### PDP 

pdp_iml = interpret.blackbox.PartialDependence(model, X_train)
pdp_global_iml = pdp_iml.explain_global()

# TODO: Check why there is a one in interpret.show(pdp...)
exp_key = 0
interpret.show(pdp_iml.explain_global(), exp_key)

### Morris Sensitivity Analysis 

# msa_iml = interpret.blackbox.MorrisSensitivity(model, X_train)
# msa_global_iml = msa_iml.explain_global()

# interpret.show(msa_global_iml, 0)

### Dashboard 

interpret.show([exp_local_shap, exp_local_lime, pdp_global_iml])
